Rootkit (Brave Browser)
WARNING -- MALICIOUS SOFTWARE;

Source: Wikepedia*

Rootkit installation can be automated (though highly unlikely, it takes govt. resources or an expert computer scientist to pull that off) or an attacker can upload it after having obtained root or admin access.*

Obtaining this access is a result of a direct attack on a system, i.e. exploiting a known vulnerability (such as privilege escalation) or a password (obtained by cracking or social engineering tactics like "phishing"). Once installed, it becomes possible to hide the intrusion as well as to maintain privileged access. Full control over a system means that existing software can be modified, including software that might otherwise be used to detect or circumvent it.*

Rootkit detection is difficult because a rootkit may be able to subvert the software that is intended to find it. Detection methods include using an alternative and trusted operating system, behavioral-based methods, signature scanning, difference scanning, and memory dump analysis. Removal can be complicated or practically impossible, especially in cases where the rootkit resides in the kernel; reinstallation of the operating system may be the only available solution to the problem. When dealing with firmware rootkits, removal may require hardware replacement or specialized equipment.*

Here we have an example of a rootkit on a user attempting to load a webpage using Brave's updated Advanced Shield web browser; thus illustrating the level of complexity and elegance needed within the rootkit in order to breach a paranoid cybersecurity freak's network/system.